Difference between = and ==
Added sentence to clear up difference between = and ==. Although this may be extraneous to the article and obvious, I am a student and I found some of my classmates making this mistake when first learning to code. And if the difference between & and && is not intuitively obvious, maybe this is also an important distinction to make.